+++ pkg/session/session.go
@@ -12,6 +12,17 @@ import (
 
 type ctxKey struct{}
 
+type storeCtxKey struct{}
+
+func WithStore(r *http.Request, s Store) *http.Request {
+	return r.WithContext(context.WithValue(r.Context(), storeCtxKey{}, s))
+}
+
+func StoreFromCtx(ctx context.Context) Store {
+	s, _ := ctx.Value(storeCtxKey{}).(Store)
+	return s
+}
+
 type Store interface {
 	Get(r *http.Request, key string) (string, error)
 	Set(w http.ResponseWriter, r *http.Request, key, value string, opts *Options) error